+++ test/operator/test_spin.py
@@ -21,6 +21,7 @@ N = 4
 
 @pytest.mark.parametrize("S", [1 / 2, 1, 3 / 2])
 def test_pauli_algebra(S):
+    N = 2
     hi = nk.hilbert.Spin(S) ** N
 
     for i in range(N):
@@ -31,18 +32,31 @@ def test_pauli_algebra(S):
         sm = spin.sigmam(hi, i)
         sp = spin.sigmap(hi, i)
 
-        assert_almost_equal(0.5 * (sx - 1j * sy).to_dense(), sm.to_dense())
-        assert_almost_equal(0.5 * (sx + 1j * sy).to_dense(), sp.to_dense())
-        assert_almost_equal(0.5 * (sx.to_dense() - 1j * sy.to_dense()), sm.to_dense())
-        assert_almost_equal(0.5 * (sx.to_dense() + 1j * sy.to_dense()), sp.to_dense())
+        assert_almost_equal(
+            np.asarray(0.5 * (sx - 1j * sy).to_dense()), np.asarray(sm.to_dense())
+        )
+        assert_almost_equal(
+            np.asarray(0.5 * (sx + 1j * sy).to_dense()), np.asarray(sp.to_dense())
+        )
+        assert_almost_equal(
+            np.asarray(0.5 * (sx.to_dense() - 1j * sy.to_dense())),
+            np.asarray(sm.to_dense()),
+        )
+        assert_almost_equal(
+            np.asarray(0.5 * (sx.to_dense() + 1j * sy.to_dense())),
+            np.asarray(sp.to_dense()),
+        )
 
         if S == 1 / 2:
             Imat = np.eye(hi.n_states)
 
             # check that -i sx sy sz = I
-            assert_almost_equal((-1j * sx @ sy @ sz).to_dense(), Imat)
             assert_almost_equal(
-                (-1j * sx.to_dense() @ sy.to_dense() @ sz.to_dense()), Imat
+                np.asarray((-1j * sx @ sy @ sz).to_dense()), np.asarray(Imat)
+            )
+            assert_almost_equal(
+                np.asarray(-1j * sx.to_dense() @ sy.to_dense() @ sz.to_dense()),
+                np.asarray(Imat),
             )